+++ CesiumITwinClient/include/CesiumITwinClient/Connection.h
@@ -13,9 +13,11 @@
 #include <CesiumAsync/Future.h>
 #include <CesiumAsync/IAssetAccessor.h>
 #include <CesiumClientCommon/OAuth2PKCE.h>
+#include <CesiumITwinClient/GeospatialFeatureCollection.h>
 #include <CesiumITwinClient/PagedList.h>
 #include <CesiumUtility/Result.h>
 #include <CesiumUtility/Uri.h>
+#include <CesiumVectorData/GeoJsonObjectTypes.h>
 
 #include <memory>
 #include <optional>
@@ -177,6 +179,32 @@ public:
   cesiumCuratedContent();
 
   /**
+   * @brief Retrieve all Collections (Feature Classes) that contain features
+   * within an iTwin.
+   *
+   * @param iTwinId The ID of the iTwin to retrieve all collections from.
+   * @see
+   * https://developer.bentley.com/apis/geospatial-features/operations/get-collections/
+   */
+  CesiumAsync::Future<
+      CesiumUtility::Result<std::vector<GeospatialFeatureCollection>>>
+  geospatialFeatureCollections(const std::string& iTwinId);
+
+  /**
+   * @brief Returns one or more pages of GeoJSON features in this iTwin.
+   *
+   * @param iTwinId The ID of the iTwin to load data from.
+   * @param collectionId The ID of the data collection to load.
+   * @param limit The maximum number of items per page, between 1 and 10,000.
+   */
+  CesiumAsync::Future<
+      CesiumUtility::Result<PagedList<CesiumVectorData::GeoJsonFeature>>>
+  geospatialFeatures(
+      const std::string& iTwinId,
+      const std::string& collectionId,
+      uint32_t limit = 10000);
+
+  /**
    * @brief Creates a new `Connection` with the provided tokens.
    *
    * It's recommended to use the \ref Connection::authorize method to create a
@@ -186,8 +214,8 @@ public:
    * @param asyncSystem The \ref CesiumAsync::AsyncSystem to use.
    * @param pAssetAccessor The \ref CesiumAsync::IAssetAccessor to use for
    * making requests to the iTwin API.
-   * @param accessToken An \ref AuthenticationToken object created from parsing
-   * the obtained iTwin access token.
+   * @param authenticationToken An \ref AuthenticationToken object created from
+   * parsing the obtained iTwin access or share token.
    * @param refreshToken A refresh token to use to fetch new access tokens as
    * needed, if any.
    * @param clientOptions The set of options to use when interacting with the
@@ -196,40 +224,44 @@ public:
   Connection(
       const CesiumAsync::AsyncSystem& asyncSystem,
       const std::shared_ptr<CesiumAsync::IAssetAccessor>& pAssetAccessor,
-      const AuthenticationToken& accessToken,
+      const AuthenticationToken& authenticationToken,
       const std::optional<std::string>& refreshToken,
       const CesiumClientCommon::OAuth2ClientOptions& clientOptions)
       : _asyncSystem(asyncSystem),
         _pAssetAccessor(pAssetAccessor),
-        _accessToken(accessToken),
+        _authenticationToken(authenticationToken),
         _refreshToken(refreshToken),
         _clientOptions(clientOptions) {}
 
   /**
    * @brief Returns the \ref AuthenticationToken object representing the parsed
-   * JWT access token.
+   * JWT access or share token.
    */
-  const AuthenticationToken& getAccessToken() const { return _accessToken; }
+  const AuthenticationToken& getAuthenticationToken() const {
+    return this->_authenticationToken;
+  }
+
   /**
-   * @brief Sets the access token that will be used for API calls.
+   * @brief Sets the access or share token that will be used for API calls.
    *
    * @param authToken The new auth token.
    */
-  void setAccessToken(const AuthenticationToken& authToken) {
-    _accessToken = authToken;
+  void setAuthenticationToken(const AuthenticationToken& authToken) {
+    this->_authenticationToken = authToken;
   }
 
   /**
    * @brief Returns the refresh token used to obtain new access tokens, if any.
    */
   const std::optional<std::string>& getRefreshToken() const {
-    return _refreshToken;
+    return this->_refreshToken;
   }
+
   /**
    * @brief Sets the refresh token used to obtain new access tokens, if any.
    */
   void setRefreshToken(const std::optional<std::string>& refreshToken) {
-    _refreshToken = refreshToken;
+    this->_refreshToken = refreshToken;
   }
 
 private:
@@ -241,13 +273,15 @@ private:
   listIModelMeshExports(const std::string& url);
   CesiumAsync::Future<CesiumUtility::Result<PagedList<ITwinRealityData>>>
   listITwinRealityData(const std::string& url);
+  CesiumAsync::Future<
+      CesiumUtility::Result<PagedList<CesiumVectorData::GeoJsonFeature>>>
+  listGeospatialFeatures(const std::string& url);
 
-  CesiumAsync::Future<CesiumUtility::Result<std::string_view>>
-  ensureValidToken();
+  CesiumAsync::Future<CesiumUtility::Result<std::string>> ensureValidToken();
 
   CesiumAsync::AsyncSystem _asyncSystem;
   std::shared_ptr<CesiumAsync::IAssetAccessor> _pAssetAccessor;
-  AuthenticationToken _accessToken;
+  AuthenticationToken _authenticationToken;
   std::optional<std::string> _refreshToken;
   CesiumClientCommon::OAuth2ClientOptions _clientOptions;
 };