+++ CesiumITwinClient/src/AuthenticationToken.cpp
@@ -14,6 +14,7 @@
 #include <string>
 #include <string_view>
 #include <utility>
+#include <variant>
 #include <vector>
 
 using namespace CesiumUtility;
@@ -64,12 +65,21 @@ AuthenticationToken::parse(const std::string& tokenStr) {
         ErrorList::error("Missing payload contents."));
   }
 
+  int64_t expired = JsonHelpers::getInt64OrDefault(json, "exp", 0);
+
+  const auto& iTwinIdValue = json.FindMember("iTwinId");
+  if (iTwinIdValue != json.MemberEnd()) {
+    return Result<AuthenticationToken>(AuthenticationToken(
+        tokenStr,
+        iTwinIdValue->value.GetString(),
+        expired));
+  }
+
   std::string name = JsonHelpers::getStringOrDefault(json, "name", "");
   std::string userName =
       JsonHelpers::getStringOrDefault(json, "preferred_username", "");
   std::vector<std::string> scopes = JsonHelpers::getStrings(json, "scope");
   int64_t notValidBefore = JsonHelpers::getInt64OrDefault(json, "nbf", 0);
-  int64_t expired = JsonHelpers::getInt64OrDefault(json, "exp", 0);
 
   return Result<AuthenticationToken>(AuthenticationToken(
       tokenStr,
@@ -86,7 +96,41 @@ bool AuthenticationToken::isValid() const {
           std::chrono::system_clock::now().time_since_epoch())
           .count();
 
-  return currentTimeSinceEpoch >= _notValidBefore &&
-         currentTimeSinceEpoch < _expires;
+  const AccessTokenContents* pAccessTokenContents =
+      std::get_if<AccessTokenContents>(&this->_contents);
+  if (pAccessTokenContents) {
+    return currentTimeSinceEpoch >= pAccessTokenContents->notValidBefore &&
+           currentTimeSinceEpoch < this->_expires;
+  }
+
+  return currentTimeSinceEpoch < this->_expires;
 }
+
+std::string AuthenticationToken::getTokenHeader() const {
+  struct TokenHeaderVisitor {
+    const std::string& tokenStr;
+    std::string operator()(const AccessTokenContents&) {
+      return fmt::format("Bearer {}", tokenStr);
+    }
+    std::string operator()(const std::string&) {
+      return fmt::format("Basic {}", tokenStr);
+    }
+  };
+
+  return std::visit(TokenHeaderVisitor{this->_token}, this->_contents);
+}
+AuthenticationToken::AuthenticationToken(
+    const std::string& token,
+    std::string&& name,
+    std::string&& userName,
+    std::vector<std::string>&& scopes,
+    int64_t notValidBefore,
+    int64_t expires)
+    : _token(token),
+      _contents(AccessTokenContents{
+          std::move(name),
+          std::move(userName),
+          std::move(scopes),
+          notValidBefore}),
+      _expires(expires) {}
 } // namespace CesiumITwinClient